moved header plugins to app
