+++ community/apps/bit-dev/app.module.scss
@@ -1,15 +1,41 @@
+@import '~@teambit/base-ui.layout.breakpoints/_breakpoints.scss';
+
 .app {
-    --bit-header-height: 40px;
+  --bit-header-height: 40px;
 }
 .label {
-    a {
-        background-color: var(--bit-bg-heavy, #f6f6f6) !important;
-        &:hover {
-            background-color: var(--bit-border-color-lightest, #ededed) !important;
-        }
+  a {
+    background-color: var(--bit-bg-heavy, #f6f6f6) !important;
+    &:hover {
+      background-color: var(--bit-border-color-lightest, #ededed) !important;
     }
+  }
 }
 
 .frame {
-    border: 1px solid var(--bit-border-color-lightest, #ededed);
+  border: 1px solid var(--bit-border-color-lightest, #ededed);
+}
+
+.inspect {
+  display: flex;
+  align-items: center;
+  padding-left: 24px;
+  > span {
+    margin-right: 4px;
+  }
+  @media screen and (max-width: $br-md) {
+    font-size: 12px;
+    padding-left: 8px;
+  }
+  @media screen and (max-width: $br-sm) {
+    display: none;
+  }
+}
+
+.githubLink {
+  @media screen and (max-width: $br-md) {
+    > span:not(:first-child) {
+      display: none;
+    }
+  }
 }
